Watch files and copy again when changed
This change is necessary in order to have RSG live update when anything changes in the components.
The idea is to have something build the components into a js and css file,
RSG will detect changes and update accordingly, and then something like browserSync can
detect changes on RSG's dist and refresh the browser.