Don't watch nor copy opts.files
Why:

Until now, the files in `opts.files` that existed in the file system,
were being copied into the styleguide folder on every run.
They were placed inside the `files` folder. A reference to all files in
`opts.files` was then inserted in the generated `index.html`.

We struggled with the fact that, whenever we change one of those
files we had to restart RSG (in order for it to be copied
into the `files` directory again). The reason for this is we use
RSG to preview a generated version of an components library (Styleguide
Driven Development?).

We run RSG alongside our library's build tool. We change
some source code, our build tool runs, then RSG runs, then the browser refreshes.

To solve this issue, we made RSG watch the local files in `opts.files`
and copy them automatically on change.

The problem is that this increased our feedback loop.

The new idea is to not copy files anymore. RSG will only insert references to
the files in `opts.files`, the rest will be taken care by other tool,
such as brwserSync. More details on this are on the README changes.

This change addresses the need by:

* Removing everything related to copying files.
* Updating the example accordingly.
* Updating the README accordingly.